fix unit tests and minor refactor
