+++ src/__mocks__/@mov-ai/mov-fe-lib-core.js
@@ -1,1 +1,3 @@
-export const Utils = {};
+export const Utils = { randomId: () => Math.floor(100 * Math.random()) };
+export const CONSTANTS = { GLOBAL_WORKSPACE: "global" }
+export const MasterDB = { subscribe: () => { } }
