--- src/models/subModels/ContainerConf/schema.js
@@ -1,5 +1,0 @@
-const Joi = require("joi");
-
-const schema = Joi.object({});
-
-export default schema;
